Final Boss Update
The final boss with the animations has been added
Rework to the last scene flow has been done.
The dialogue system references were updated. Now the dialogue events are only started from the actors using the Dialogue Actions scripts as a connection with the rest of the assets in the scene to allow a proper flow
The last scene was also updated to be able to interact properly with the dialogue system changes