BreakableObjects Cesar Update
The initial image was changed in the Sprite Section in the Sprite Renderer component.
The mechanic works fine.
The sound must be updated
The colliders must be updated. (The colliders must be updated to adapt to the new sprites)
The sprites must be update(They must have the same canvas to be replaced without changing position
The positions in the scenes must be updated accordingly to the new sprites. (All scenes that have this items must be updated since most of them will be in weird positions)