SavingSystemRework V0.1
The Easy Save Asset was added to the game.
It was implemented the use of it in all assets and components in Level1 and Splash Screens.
Remains to add it to the Dialogue Systems and the rest of the game.
The work is that all variables are updated at the Player Data FSM (Those that are neede to save at specific moments). Then when a save process is needed the system will save all the variables with the current value. Then when the game is started again the system will load back all data to the Player Data for those Assets that need them to have the value at their disposal